requirements.txt
asgiref==3.3.4
Django==3.2.2
django-classy-tags==2.0.0
django-cms==3.8.0
django-emailtemplates==1.1.2
django-formtools==2.3
django-sekizai==2.0.0
django-treebeard==4.5.1
djangocms-admin-style==2.0.2
djangorestframework==3.12.4
mysqlclient==2.0.3
pytz==2021.1
sqlparse==0.4.1
